requirements.txt
streamlit==1.39.0
llama-index==0.11.0
llama-index-llms-groq==0.2.0
llama-index-embeddings-cohere==0.3.0
pypdf==4.3.0
python-docx==1.1.2
docx2txt==0.8
pillow==10.4.0
python-dotenv==1.0.1